CMakeDetermine(C|CXX)Compiler: Prefer generic system compilers
Teach CMake to prefer the system default compiler automatically.  Look
for system compilers with generic names before GNU-specific names.
Other vendor-specific names already have lower priority.

 C:   Prefer "cc" before GNU "gcc"
 C++: Prefer "CC" before "c++" or GNU "g++"

This will change the default compiler selection for existing build
scripts that do not specify a compiler when run on machines with
separate system and GNU compilers both installed in the PATH.

We do not make this change in default behavior lightly.  Howerver:

(1) If a given build really needs specific compilers one should specify
    them explicitly e.g. by setting CC and CXX in the environment.

(2) The motivating case is to prefer the system Clang on OS X machines
    over the older GNU compilers typically also installed.  On newer
    OS X versions the name "c++" links to Clang so the old behavior
    picked Clang for C++ and GNU for C by default.

(3) Other than the motivating OS X case the conditions under which the
    behavior changes do not tend to exist in default OS installations.
    They typically occur only on non-GNU systems with manually-installed
    GNU compilers.

(4) The consequences of the new behavior are not dire.  At worst the
    project fails to compile with the system compiler when it previously
    worked with the non-system GNU compiler.  Such failure is easy to
    work around (see #1).

In short this change fixes default behavior on a widely-used platform
at the cost of a modest change in behavior in less-common conditions.