Revert "CMakeDetermine(C|CXX)Compiler: Prefer generic system compilers"
This reverts commit 27b74445b49cee310ad0951b1cb3bdb0645b95f4.

On case-insensitive OS X filesystems using "CC" as the C++ compiler
acts as a C compiler when generating the link line and does not add
the -lstdc++ library.  However, if we prefer "cc" for C and "c++"
then on platforms with non-GNU system compilers named "cc" and "CC"
and a GNU compiler installed named "c++" we will get mixed-vendor
selection.  Another solution will be needed to avoid mixing old GNU
and new Clang on OS X.